+++ packages/transaction-recovery/src/evm-add-native-gas/client.ts
@@ -3,8 +3,8 @@ import { createAxelarConfigClient } from "@axelarjs/api/axelar-config";
 import { createAxelarQueryClient } from "@axelarjs/api/axelar-query";
 import { createGMPClient } from "@axelarjs/api/gmp";
 
-import type { EvmAddNativeGasParams } from "../types";
 import { addNativeGas as baseAddNativeGas } from "./isomorphic";
+import type { EvmAddNativeGasParams } from "./types";
 
 /**
  * Send additional native gas to the source transaction to cover the insufficient gas fee.
@@ -16,7 +16,7 @@ import { addNativeGas as baseAddNativeGas } from "./isomorphic";
  * - `evmSendOptions`: The options to send the transaction.
  * @returns The native gas payment transaction.
  */
-export function evmAddNativeGas(params: EvmAddNativeGasParams) {
+export function addNativeGasEvm(params: EvmAddNativeGasParams) {
   const { environment } = params;
 
   return baseAddNativeGas(params, {
@@ -27,4 +27,4 @@ export function evmAddNativeGas(params: EvmAddNativeGasParams) {
   });
 }
 
-export default evmAddNativeGas;
+export default addNativeGasEvm;
