+++ packages/transaction-recovery/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./add-gas";
+export * from "./cosmos-add-gas";
 export * from "./evm-add-native-gas";
